Give the Tcl interpreter full power by calling Tcl_Init.
Source in the file syslog.tcl that defines a proc named syslog.

Set elements of a Tcl array named "message": clock, timestamp, facility,
priority, program, kernel (sometimes), and msg.

Use Tcl_Eval to invoke the proc named syslog when messages are to be logged.

If syslog debugging is enabled and Tcl interpreter initialization fails or
calls to Tcl to set variables or to Tcl_Eval fail, debug messages are emitted
including the Tcl error message.